write addNewInput Func
